Integrations for #119 and #120
- VaultConfigurationClient.cs (both): Remove direct check for target frameworks within code.
- *.settings: Migrate to LTS
- SystemGlobal.cs: Clean up some code.
- PolyMath.cs: Update NETFRAMEWORK check.
- EventLogLogger.cs: Use new atomic impl and remove the windows marking for the class (unsupported)
- ExponentialBackoff.cs: Change from byte to int.
- Atomic.cs: Clean up a lot of code and use generics mixed with the "dynamic" type. (TODO: Guard against non-number types)
- dispatcher.cs and interleave.cs: Update target framework integ.
- CommandRegistry.cs: Use new atomic impl
- *WorkQueue.cs: Update atomic impl and change around NETFRAMEWORK references.
- GridUtility: Change around atomic references, change around framework references and change base url from sitetest4.robloxlabs.com to roblox.com
- Remove Unit Tests
- Add targets files

+++ Assemblies/MFDLabs.Sentinels/Backoff/ExponentialBackoff.cs
@@ -5,10 +5,10 @@ namespace MFDLabs.Sentinels
 {
     public static class ExponentialBackoff
     {
-        public static TimeSpan CalculateBackoff(byte attempt, byte maxAttempts, TimeSpan baseDelay, TimeSpan maxDelay, Jitter jitter = Jitter.None) 
+        public static TimeSpan CalculateBackoff(uint attempt, uint maxAttempts, TimeSpan baseDelay, TimeSpan maxDelay, Jitter jitter = Jitter.None) 
             => CalculateBackoff(attempt, maxAttempts, baseDelay, maxDelay, jitter, () => Random.NextDouble());
 
-        private static TimeSpan CalculateBackoff(byte attempt, byte maxAttempts, TimeSpan baseDelay, TimeSpan maxDelay, Jitter jitter, Func<double> nextRandomDouble)
+        private static TimeSpan CalculateBackoff(uint attempt, uint maxAttempts, TimeSpan baseDelay, TimeSpan maxDelay, Jitter jitter, Func<double> nextRandomDouble)
         {
             if (maxAttempts > CeilingForMaxAttempts) 
                 throw new ArgumentOutOfRangeException($"{maxAttempts} must be less than or equal to {CeilingForMaxAttempts}");
@@ -25,7 +25,7 @@ namespace MFDLabs.Sentinels
             return delay;
         }
 
-        private const byte CeilingForMaxAttempts = 10;
+        private const uint CeilingForMaxAttempts = 10;
         private static readonly ThreadLocalRandom Random = new ThreadLocalRandom();
     }
 }
